requirements.txt
streamlit
langchain
langchain-community
langchain-core
langchain-groq
langchain-huggingface
youtube-transcript-api
faiss-cpu
sentence-transformers
python-dotenv
